+++ lib/ooxml_parser/xlsx_parser/xlsx_data/view_model/workbook/pivot_cache/pivot_cache_definition/cache_fields/cache_field.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module OoxmlParser
+  # Class for parsing <cacheField> tag
+  class CacheField < OOXMLDocumentObject
+    # Parse `<worksheetSource>` tag
+    # # @param [Nokogiri::XML:Element] node with WorksheetSource data
+    # @return [WorksheetSource]
+    def parse(node)
+      node.attributes.each do |key, value|
+      end
+      self
+    end
+  end
+end